+++ tibiakt-core/src/main/kotlin/com/galarzaa/tibiakt/core/section/library/creature/model/BossEntry.kt
@@ -28,4 +28,4 @@ import kotlinx.serialization.Serializable
 public data class BossEntry(
     override val name: String,
     override val identifier: String,
-) : BaseCreatureEntry
+) : BaseCreature