+++ tibiakt-core/src/main/kotlin/com/galarzaa/tibiakt/core/section/library/creature/model/CreatureEntry.kt
@@ -29,7 +29,7 @@ import kotlinx.serialization.Serializable
 public data class CreatureEntry(
     override val name: String,
     override val identifier: String,
-) : BaseCreatureEntry {
+) : BaseCreature {
     /**
      * The URL to the creature's page.
      */